Move BerryBush to util subpackage
